+++ flumine/config.py
@@ -18,3 +18,9 @@ raise_errors = False  # used for call_check_market / call_process_market_book
 max_execution_workers = 32  # max number of workers in execution thread pool
 
 async_place_orders = False  # async place orders
+
+# latencies used for backtesting
+place_latency = 0.120
+cancel_latency = 0.170
+update_latency = 0.150
+replace_latency = 0.280